singleDoker/dj_aws_template/requirements.txt
Django==3.2.12
djangorestframework==3.12.4
django-cors-headers==3.10.0


psycopg2==2.8.2
django-storages==1.7.1
django-redis==4.12.1
PynamoDB~=5.1
botocore~=1.20.47
boto3==1.17.112

Cython==0.29.27



